Rename route and remove required icon on speaker social
